refactor shared code into ci_base
